+++ src/js/components/Skeleton/index.d.ts
@@ -1,0 +1,9 @@
+import * as React from 'react';
+
+export interface SkeletonProps {
+  kind?: string;
+}
+
+declare const Skeleton: React.FC<SkeletonProps>;
+
+export { Skeleton };
