+++ src/js/components/Skeleton/propTypes.js
@@ -1,0 +1,9 @@
+import PropTypes from 'prop-types';
+
+let PropType = {};
+if (process.env.NODE_ENV !== 'production') {
+  PropType = {
+    kind: PropTypes.string,
+  };
+}
+export const SkeletonPropTypes = PropType;
